+++ src/library/DIPS.Mobile.UI/Components/Buttons/dotnet/ButtonHandler.cs
@@ -1,0 +1,14 @@
+namespace DIPS.Mobile.UI.Components.Buttons;
+
+public partial class ButtonHandler : Microsoft.Maui.Handlers.ButtonHandler
+{
+    private partial void AppendPropertyMapper()
+    {
+        
+    }
+    
+    private static partial void MapAdditionalHitBoxSize(ButtonHandler handler, Button button)
+    {
+        
+    }
+}