+++ src/library/DIPS.Mobile.UI/Components/Images/ImageButton/dotnet/ImageButtonHandler.cs
@@ -6,4 +6,8 @@ public partial class ImageButtonHandler
     {
     }
 
+    private static partial void MapAdditionalHitBoxSize(ImageButtonHandler handler, ImageButton imageButton)
+    {
+    }
+    
 }