+++ src/library/DIPS.Mobile.UI/Platforms/Android/ButtonWithExtraTappableArea.cs
@@ -1,0 +1,13 @@
+using Android.Content;
+using Google.Android.Material.Button;
+using Button = Android.Widget.Button;
+
+namespace DIPS.Mobile.UI.Platforms.Android;
+
+internal class ButtonWithExtraTappableArea : MaterialButton
+{
+    public ButtonWithExtraTappableArea(Context context) : base(context)
+    {
+        
+    }
+}